requirements.txt
streamlit
python-dotenv
requests
google-cloud-storage
google-auth
openai